Refactor ft_printf functions and add get_next_line implementation

+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: vokotera <vokotera@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/27 15:48:15 by vokotera          #+#    #+#             */
-/*   Updated: 2025/11/27 16:35:27 by vokotera         ###   ########.fr       */
+/*   Updated: 2025/12/01 16:14:18 by vokotera         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,6 @@
 # define FT_PRINTF_H
 
 # include <stdarg.h>
-# include <stdio.h>
 # include <unistd.h>
 # include <stdint.h> 
 
@@ -22,8 +21,8 @@ int	printchar(int c);
 int	ft_printf(const char *format, ...);
 int	printstr(const char *s);
 int	printptr(void *ptr);
-int tohex(uintptr_t decimal, int uppercase);
-int	printdecimal(long n);
-int	print_unsigned(unsigned long n);
+int	tohex(uintptr_t decimal, int uppercase);
+int	printdecimal(int n);
+int	print_unsigned(unsigned int n);
 
 #endif
